Removing dependency on react-native-windows
At best, this should have a peer dependency on react-native-windows (in the manner that [other packages](https://github.com/kmagiera/react-native-gesture-handler) do for react-native). In the past, I've added react-native-windows even as a dev dependency for compatible modules.  However, it absolutely should not be a strict dependency (much in the same way that react-native is not a dependency here.