reverse order of operations in onHostPause()
Upon pausing the video by backgrounding the app the video would restart upon bringing it to the foreground.  This was due to using mActiveStatePauseStatus which is false after pausing via onHostPause().  This behavior is undesirable for our media playing app.  An alternative to this would have been to use mPaused as the boolean in onHostResume().