Add ability to save streamed video
I have only implemented iOS part since I am no good with Java for such a task. If onVideoSaved is set in props video is exported to documents directory and the path is returned as a string. If the callback prop is not set the video is not exported. This helps developers prevent downloading the same data multiple times.