Fixes bug that causes player to crash when fullscreening at initialization (#1234)
Beause the `setSrc` method now runs async, it is possible for the player to not be initialized at the point a user calls the `setFullscreen` method via `presentFullscreenPlayer`.
A good example of this is when a user of the library wants to fullscreen a video as soon as the ref for the video component is captured. To prevent this problem, in the case where the user is trying to fullscreen the video but the player has not yet been instantiated, this will set a flag letting the system know that the user would like to fullscreen the video as soon as the player is initialized. Upon initialization, that flag will be read, the player will be fullscreened, and the flag will be set back.